+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - mutation on read-only futures (`completeWith`, `orTimeout`) is not prevented
 - `ExtendedFuture#cancel(true)` on non-interruptible downstream stage does not interrupt upstream
 - some delegation methods are missing on internal `ExtendedFuture#WrappingFuture` class
-- Combiner#toMap ignores null results from futures
+- `Combiner#toMap` ignores null results from futures
+- ClassCastException when re-wrapping interruptible futures via `asCancellableByDependents`, `withInterruptibleStages`, `withDefaultExecutor`
 
 
 ## [1.1.1] 2025-05-01